+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
-
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+- Update password style to a single uppercase and number
+- The generated password is now shown as an alfred result
+- `max_repeats` default parameter set to 1
 
+### Removed
+- Notifications are removed.
+
+## [1.0.0] - 2020-02-05
 ### Added
 - Password generation logic
 - Variables to control how the password is generated
 
-[Unreleased]: https://github.com/erremauro/alfred-random-password
+[Unreleased]: https://github.com/erremauro/alfred-random-password/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/erremauro/alfred-random-password/releases/tag/v1.0.0